Partial bug fix for delete assocs for has_one
Correct whitelist_classes() to return also has_one names.

Try delete if delete_all is not accepted: should work for has_one
associations w/o additional associations.

TODO: fix for bedgraph_files, which is not only has_one, biut also has
associations with per transcript coverages, so simple delete does not
work. Need to delete also per transcript coverages records, so user
should not be able to delete just bedgraph_files alone.

+++ app/controllers/entity_controller.rb
@@ -5,15 +5,30 @@ class EntityController < ApplicationController
     read_group_id = entity_params[:read_group_id]
     read_group = ReadGroup.find(read_group_id)
 
+    puts "whitelist_classes=#{whitelist_classes.inspect};"
     allowed_associations = entity_params[:associations].select do |association|
       whitelist_classes.include?(association.to_sym)
     end
 
+    # TODO 1: change to del/destroy all assoc records; depending on the model; maybe use only 1 master button instead of many for bedgraph and per trx cov; 03/02/16 W;
+    status = 501
     allowed_associations.each do |association|
-      if read_group.send(association).delete_all
-        status = 200
+
+      if read_group.send(association).respond_to?(:delete_all)
+        if read_group.send(association).delete_all
+          status = 200
+        else
+          status = 500
+        end
+      elsif read_group.send(association).respond_to?(:delete)
+        if read_group.send(association).delete
+          status = 200
+        else
+          status = 500
+        end
       else
-        status = 500
+        raise("Cannot delete association=#{association.inspect} " +
+              "for read_group=#{read_group.inspect}")
       end
     end
 
@@ -23,7 +38,8 @@ class EntityController < ApplicationController
   private
 
   def whitelist_classes
-    ReadGroup.reflect_on_all_associations(:has_many).map(&:name)
+    (ReadGroup.reflect_on_all_associations(:has_many).map(&:name) +
+     ReadGroup.reflect_on_all_associations(:has_one).map(&:name))
   end
 
   def entity_params
